move test classes back into the generator project
